requirements.txt
streamlit>=1.36.0
google-api-python-client>=2.129.0
google-auth>=2.31.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
tenacity>=8.2.3
python-dateutil>=2.9.0
rich>=13.7.1
pandas>=2.2.2
humanize>=4.9.0
psutil>=5.9.8
pytest>=8.3.1
fpdf2>=2.7.9
